prospect_web_application/requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
boto3==1.11.11
botocore==1.14.11
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Django==2.2.10
docutils==0.15.2
idna==2.8
jmespath==0.9.4
numpy==1.18.1
pandas==1.0.1
pkg-resources==0.0.0
psycopg2==2.8.4
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
requests==2.22.0
s3transfer==0.3.2
selenium==3.141.0
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.13
sqlparse==0.3.0
urllib3==1.25.8

